Header: outline analyzing routines
